+++ components/Map.tsx
@@ -14,8 +14,6 @@ export default function Map({ artists }: MapProps) {
     iconSize: [38, 95],
     iconAnchor: [22, 94],
     popupAnchor: [-3, -76],
-    shadowSize: [68, 95],
-    shadowAnchor: [22, 94],
   });
 
   return (
