+++ tsconfig.json
@@ -22,7 +22,7 @@
     "globalStyles.js",
     "pages/api/index.ts",
     "db/Artist.js",
-    "lib/upload.ts",
+    "lib/uploadUserData.ts",
     "pages/map.tsx"
   ],
   "exclude": ["node_modules", "codesnippets"]
